requirements.txt
gizeh==0.1.10
PyYAML==3.13
tensorflow-gpu==1.0.1
